chartlib: refactor chart to an abstract class
