+++ src/ipyvizzu/__init__.py
@@ -1,6 +1,7 @@
 """Jupyter notebook integration of Vizzu."""
 
 from .jupyter.chart import Chart
+from .jupyter.template import DisplayTemplate
 from .chartlib.animation import (
     Animation,
     PlainAnimation,
@@ -13,5 +14,5 @@ from .chartlib.animation import (
 )
 from .chartlib.method import Method, Animate, Feature, Store, EventOn, EventOff, Log
 from .chartlib.json import RawJavaScript, RawJavaScriptEncoder
-from .chartlib.template import ChartProperty, DisplayTarget, DisplayTemplate
+from .chartlib.template import ChartProperty, DisplayTarget
 from .chartlib.event import EventHandler
