+++ src/ipyvizzu/chartlib/template.py
@@ -3,6 +3,9 @@
 from enum import Enum
 
 
+VIZZU = "https://cdn.jsdelivr.net/npm/vizzu@~0.4.0/dist/vizzu.min.js"
+
+
 class ChartProperty(Enum):
     """An enum class for storing chart properties."""
 
@@ -23,40 +26,3 @@ class DisplayTarget(Enum):
     END = "end"
     ACTUAL = "actual"
     MANUAL = "manual"
-
-
-class DisplayTemplate:
-    """A class for storing string templates to generate javascript snippets."""
-
-    # pylint: disable=too-few-public-methods
-
-    IPYVIZZUJS = "{ipyvizzujs}"
-
-    INIT = (
-        "window.ipyvizzu.createChart(element, "
-        + "'{chart_id}', '{vizzu}', '{div_width}', '{div_height}');"
-    )
-
-    ANIMATE = (
-        "window.ipyvizzu.animate(element, "
-        + "'{chart_id}', '{display_target}', {scroll}, {chart_target}, {chart_anim_opts});"
-    )
-
-    FEATURE = "window.ipyvizzu.feature(element, '{chart_id}', '{name}', {enabled});"
-
-    STORE = "window.ipyvizzu.store(element, '{chart_id}', '{id}');"
-
-    SET_EVENT = (
-        "window.ipyvizzu.setEvent(element, "
-        + "'{chart_id}', '{id}', '{event}', event => {{ {handler} }});"
-    )
-
-    CLEAR_EVENT = (
-        "window.ipyvizzu.clearEvent(element, '{chart_id}', '{id}', '{event}');"
-    )
-
-    LOG = "window.ipyvizzu.log(element, '{chart_id}', '{chart_property}');"
-
-    CLEAR_INHIBITSCROLL = (
-        "if (window.IpyVizzu) { window.IpyVizzu.clearInhibitScroll(element); }"
-    )
